requirements.txt
fastapi==0.108.0
openai<1.3,~=1.2
pyautogen==0.2.1
pydantic==1.10.8
uvicorn
jinja2
websockets
